Split up rev module
